switch story to TS, add knobs
